+++ tests/Infrastructure.Tests/PowerShellScripts/PowerShellCommand.cs
@@ -3,6 +3,7 @@
 
 using System.Diagnostics;
 using System.Runtime.InteropServices;
+using Aspire.TestUtilities;
 using Xunit;
 
 namespace Infrastructure.Tests;
@@ -293,38 +294,5 @@ public class CommandException : Exception
     public CommandException(string message, CommandResult result) : base(message)
     {
         Result = result;
-    }
-}
-
-/// <summary>
-/// Extension methods for Process handling.
-/// </summary>
-internal static class ProcessExtensions
-{
-    public static bool TryGetHasExited(this Process process)
-    {
-        try
-        {
-            return process.HasExited;
-        }
-        catch (InvalidOperationException ie) when (ie.Message.Contains("No process is associated with this object"))
-        {
-            return true;
-        }
-    }
-
-    public static void CloseAndKillProcessIfRunning(this Process? process)
-    {
-        if (process is null || process.TryGetHasExited())
-        {
-            return;
-        }
-
-        if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-        {
-            process.CloseMainWindow();
-        }
-        process.Kill(entireProcessTree: true);
-        process.Dispose();
     }
 }
