+++ tests/Shared/TemplatesTesting/CommandResult.cs
@@ -8,20 +8,13 @@ using System.Text;
 namespace Aspire.Templates.Tests;
 
 // taken from https://github.com/dotnet/arcade/blob/main/src/Common/Microsoft.Arcade.Common/CommandResult.cs
-public struct CommandResult
+public struct CommandResult(ProcessStartInfo startInfo, int exitCode, string output)
 {
     public static readonly CommandResult Empty;
 
-    public ProcessStartInfo StartInfo { get; }
-    public int ExitCode { get; }
-    public string Output { get; }
-
-    public CommandResult(ProcessStartInfo startInfo, int exitCode, string output)
-    {
-        StartInfo = startInfo;
-        ExitCode = exitCode;
-        Output = output;
-    }
+    public ProcessStartInfo StartInfo { get; } = startInfo;
+    public int ExitCode { get; } = exitCode;
+    public string Output { get; } = output;
 
     public CommandResult EnsureSuccessful(string messagePrefix = "", bool suppressOutput = false)
         => EnsureExitCode(0, messagePrefix, suppressOutput);
